调整，抽取公共方法
